Fix tests for pulling github commits.
Snapshot is a bad idea here, as commits are dynamic and change - every time it occurs snapshot will throw error. Besides snapshots for these repsonses are very big. To tackle it a new simple helper test has been introduced to check if the response is a list and has all needed fields.