Writing tests for GitLab Rest Engine' getting repos by phrase (from bottom logic to upper).
This tests encouraged to move get_repos from superclass to subclass as more and more features start to differ between GitHub and GitLab (writing tests helped find it out). Main differences are with language filter - GitLab searching engine does not support it, so it has to be filtered at the end from table. Other difference deals with contributors - GitLab GraphQL repos engine does not support it (at least not easily for the time-being), so they have to be added at the end with the use of Rest.